mood-snap-studio-main/backend/requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
python-dotenv
fastapi
uvicorn
python-multipart
Pillow
# NEW: Open Source Image Engine Tools
opencv-python-headless
rembg
huggingface_hub
numpy
scikit-image
requests
